refactored table storage and card creation into utility files
