+++ .env.example
@@ -1,6 +1,7 @@
-AZURE_STORAGE=
 APP_ID=
 APP_PASS=
-EMAIL_PASS=
 LUIS_ID=
 LUIS_KEY=
+EMAIL_PASS=
+EMAIL_USERNAME=
+AZURE_STORAGE=
